requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-BCrypt==1.0.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
pandas==2.0.0
openpyxl==3.10.0
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.0
itsdangerous==2.1.2
click==8.1.3
